Some structure around sub commands.
Commands are given the 'api' for ECM to pluck content from
and the arguments fed via the command line.  Currently the
arg parsing is centralized in main.py but that might move
into the commands folder for each command to manage.

Logs is basically funcational in grabbing some content.  Very
no frills for the moment.