Centralize html.unescape to the main api parser.
Instead of doing html.unesacpe in each place where I think it may
have html characters I just unescape the entire data stream at the
api level.  This may be too big a net but for now it seems to work
and reduces the code smell quite a bit.

I also removed expansions of the initial routers resource get_pager
call which should faciliate running commands in accounts with large
numbers of routers.  My tests generally take place in my very small
sandbox in prod which only has 3 devices.  Observations of an ECM
dev using the tool in large accounts revealed it just bogs down to
the point of being unusable.  This may help with that.