Implement proper command argument stack.
I discovered that when using multiple levels of commands the behavior
of argparse's set_defaults will simply clobber the previous values.
This is normal for single level argparsing, so I needed to rethink how
I store these command arguments when there are N+1 of them.  Now I build
a stack of them and use a scanner method to look for all the 'levels' of
commands.  It's slightly magical but the code for parsing subcommands is
actually a little cleander now.