Checkpoint for major cleanup and 'remote' cmd.
Trying to get everything using the same argument standards so tab
completion works consistently.  Also working to make sure commands use
the shellish.Table and Tree libraries in a way that lets them spool
through large datasets effectively.

Dropping the notion of configuration alltogether.  It's way to
contriersial to mess with and Jen convinced me that fire and forget is a
viable strategy if you can apply it to selections of devices.  So now
everything is through ecm-remote and it can do basic CRUD commands using
selection criteria of the users choosing.  You can apply config to a
product or firmware, or group, etc.  It's up to them.