Workaround for 'remote' api paging, csv format, file stuffs.
Use a custom pager routine for remote api calls since the pager is not
fully baked on the server side and due to being a blocking application
is is often better to do small requests when the request may not work at
all.

Better csv formatting.

Better handling in many places where we support optional file output or
redirection to pipes, etc.

Support for --repeat mode in the remote-get command.  This is a fun one.