Async remote command experiment!
This is a fun commit, this could make the remote command super powerful
for large installs.  I'm using the async http client form tornado to do
the remote API calls so it can basically firehose them out for each
router and handle each response as it becomes available.  Very
experimental.