Update to add_file_argument context style.
The add_file_argument method requires you to use a context statement
(with) on the return value.  Updated the commands using this method to
use that but also to use the original default='-' init of the arg which
provides sys.stdout as the default value when not given by the user.
This simplifies the setup a bit.