Change remote to yield 1 item per device.
The remote method from api would yield a row for each path/value pair it
found in the results which was an experement of mine.  I wasn't sure if
that was the best idea and it turns out I don't like it, so this makes
the results 1:1 with value:router even when using glob patterns that may
return more than one result per router.

Fixed some cruft around the signature for this method too.